Nydus: refactor build workflow with packer
Introduce a new package called packer, which implements the build workflow
of nydus image, as well as the export and import of build cache, the build
workflow previously implemented by the nydus driver has been wrapped into
this package.

The reason for introducing packer is that nydus intends to integrate packer
as a generic component into other build toolkits such as buildkit to more
easily support building and exporting of nydus image.

Unlike other compression format (gzip, estargz etc.) in OCI image, nydus is
divided into two types of layer, blob and bootstrap, where blob serves as
the data part of each layer of the image, and bootstrap serves as the metadata
of the whole image, the bootstrap is equivalent to the view of the whole image
filesystem after all layers overlay. For example, for an OCI image with 3 layers,
the corresponding nydus image is 4 layers (3 layers of blob + 1 layer of bootstrap),
and since nydus supports chunk-level data de-duplication, it can do inter-layer
data de-duplication and inter-image data de-duplication, so it is possible that
an OCI layer does not have a corresponding nydus blob layer, because it was
de-duplicated during the build workflow.

The packer also implements support for build cache, which is used to speed up
the conversion or build of nydus image, i.e. when a built image layer appears
in a new image, it does not need to be built again. nydus' build cache principle
is to write the blob and bootstrap data of each layer to the content store at
each build, and store the cache record. In the next build, will find the layers
that have already been built from the cache record, skip the build of these cached
layer, and reuse the cached bootstrap as parent to build a new nydus image with
new layers.

Signed-off-by: Yan Song <imeoer@linux.alibaba.com>